third_party/WebKit/LayoutTests/fullscreen/full-screen-enter-leave-enter-expected.txt

EXPECTED (document.webkitIsFullScreen == 'false') OK
EXPECTED (document.webkitCurrentFullScreenElement == 'undefined') OK
EXPECTED (document.webkitIsFullScreen == 'false') OK
EXPECTED (document.webkitCurrentFullScreenElement == 'undefined') OK
END OF TEST
